break large file into smaller files
