+++ index.js
@@ -1,2 +1,4 @@
 import { Neu3D } from './src/neu3d';
+import './src/object_loader.js';
+import './src/import_export.js';
 export { Neu3D };
